Create worker renderer for DB and file io. Communicate with renderer via socket.io.
