+++ package.json
@@ -177,6 +177,8 @@
     "@types/react-router": "^5.1.12",
     "@types/react-router-dom": "^5.1.6",
     "@types/react-test-renderer": "^16.9.3",
+    "@types/socket.io-client": "^1.4.36",
+    "@types/uuid": "^8.3.0",
     "@types/webpack-env": "^1.15.2",
     "@typescript-eslint/eslint-plugin": "^4.8.1",
     "@typescript-eslint/parser": "^4.8.1",
@@ -248,11 +250,14 @@
     "electron-updater": "^4.3.4",
     "fast-glob": "^3.2.5",
     "history": "^5.0.0",
+    "query-string": "^6.14.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "regenerator-runtime": "^0.13.5",
-    "source-map-support": "^0.5.19"
+    "socket.io-client": "^4.0.0",
+    "source-map-support": "^0.5.19",
+    "uuid": "^8.3.2"
   },
   "resolutions": {
     "@types/react": "^17.0.3"
