+++ src/__tests__/App.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '@testing-library/jest-dom';
 import { render } from '@testing-library/react';
-import App from '../App';
+import App from '../renderer/App';
 
 describe('App', () => {
   it('should render', () => {
