--- src/entities/BaseEntity.ts
@@ -1,15 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import { PrimaryKey } from '@mikro-orm/core'
-
-export abstract class BaseEntity {
-
-  @PrimaryKey({ type: 'number' })
-  id!: number;
-
-  // @Property()
-  // createdAt: Date = new Date();
-
-  // @Property({ onUpdate: () => new Date() })
-  // updatedAt: Date = new Date();
-
-}
