+++ src/main/ipc_channels/index.ts
@@ -1,0 +1,2 @@
+export * from './AppPath';
+export * from './Dialog';
