+++ src/package.json
@@ -1,7 +1,7 @@
 {
   "name": "amst",
   "productName": "amst",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "description": "A file database and tagging system.",
   "main": "./main.prod.js",
   "author": {
@@ -18,10 +18,12 @@
     "@mikro-orm/core": "^4.4.4",
     "@mikro-orm/migrations": "^4.4.4",
     "@mikro-orm/sqlite": "^4.4.4",
-    "reflect-metadata": "^0.1.13"
+    "reflect-metadata": "^0.1.13",
+    "socket.io": "^4.0.0"
   },
   "devDependencies": {
-    "@mikro-orm/cli": "^4.4.4"
+    "@mikro-orm/cli": "^4.4.4",
+    "@types/socket.io": "^2.1.13"
   },
   "mikro-orm": {
     "useTsNode": true,
