+++ src/renderer/components/File/List.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import { FileStub } from '../../../db/entities';
+
+interface FileListProps
+{
+  files: FileStub[];
+}
+
+export default function FileList({ files }: FileListProps)
+{
+  return (
+    <ul>
+      {files.map((file) =>
+      <li key={file.id}>
+        {file.name}, {file.extension}, {file.fullPath}
+      </li>)}
+    </ul>
+  );
+}
