--- src/utils/ipc/channels/index.ts
@@ -1,2 +1,0 @@
-export * from './AppPathChannel';
-export * from './DialogChannel';
